backend/requirements.txt
# SignalOS Backend Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Async Support
asyncio
aiofiles==23.2.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.0.3

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# AI/ML (Optional - for future LLM integration)
# transformers==4.35.2
# torch==2.1.1

# Image Processing (for OCR)
pillow==10.1.0
# easyocr==1.7.0  # Uncomment if needed

# System Monitoring
psutil==5.9.6

# Utilities
python-dotenv==1.0.0
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing